Implemented Basic FSM for Fish Enemy
Currently, the Fish Enemy AI has three states:
- Idle State
- Patrol State
- Chase State

The Idle state and Patrol state have the same animations, but different speeds to emphasize that the AI is in different states. While the Chase state has a completely different animation.

There is still some work needed to be done for making the detection better, but I think this current version of the FSM suffices for the A2 deliverable.